estacionaPE/backend/requirements.txt
Flask
flask-smorest
flask-sqlalchemy
flask-jwt-extended
flask-migrate
marshmallow
pymysql
python-dotenv
authlib
requests